+++ IntelligenceHub.Controllers/ToolController.cs
@@ -46,7 +46,8 @@ namespace IntelligenceHub.Controllers
         {
             try
             {
-                await GetTenantIdAsync();
+                var tenantResult = await SetUserTenantContextAsync();
+                if (!tenantResult.IsSuccess) return StatusCode(StatusCodes.Status500InternalServerError, tenantResult.ErrorMessage);
                 if (string.IsNullOrEmpty(name)) return BadRequest("Invalid request. Please check the route parameter for the profile name.");
                 var response = await _profileLogic.GetTool(name);
 
@@ -75,7 +76,8 @@ namespace IntelligenceHub.Controllers
         {
             try
             {
-                await GetTenantIdAsync();
+                var tenantResult = await SetUserTenantContextAsync();
+                if (!tenantResult.IsSuccess) return StatusCode(StatusCodes.Status500InternalServerError, tenantResult.ErrorMessage);
                 if (page < 1) return BadRequest("The page must be greater than 0.");
                 if (count < 1) return BadRequest("The count must be greater than 0.");
                 var tools = await _profileLogic.GetAllTools(page, count);
@@ -102,7 +104,8 @@ namespace IntelligenceHub.Controllers
         {
             try
             {
-                await GetTenantIdAsync();
+                var tenantResult = await SetUserTenantContextAsync();
+                if (!tenantResult.IsSuccess) return StatusCode(StatusCodes.Status500InternalServerError, tenantResult.ErrorMessage);
                 if (string.IsNullOrEmpty(name)) return BadRequest("Invalid request. Please check the route parameter for the profile name.");
                 var response = await _profileLogic.GetToolProfileAssociations(name);
                 if (response.StatusCode == APIResponseStatusCodes.NotFound) return NotFound(response.ErrorMessage);
@@ -129,7 +132,8 @@ namespace IntelligenceHub.Controllers
         {
             try
             {
-                await GetTenantIdAsync();
+                var tenantResult = await SetUserTenantContextAsync();
+                if (!tenantResult.IsSuccess) return StatusCode(StatusCodes.Status500InternalServerError, tenantResult.ErrorMessage);
                 var response = await _profileLogic.CreateOrUpdateTools(toolList);
                 if (!response.IsSuccess) return BadRequest(response.ErrorMessage);
                 else return Ok(toolList);
@@ -157,7 +161,8 @@ namespace IntelligenceHub.Controllers
         {
             try
             {
-                await GetTenantIdAsync();
+                var tenantResult = await SetUserTenantContextAsync();
+                if (!tenantResult.IsSuccess) return StatusCode(StatusCodes.Status500InternalServerError, tenantResult.ErrorMessage);
                 if (string.IsNullOrEmpty(name)) return BadRequest("Invalid request. Please check the route parameter for the profile name.");
                 if (profiles == null || profiles.Count < 1) return BadRequest("Invalid request. 'Profiles' property cannot be null or empty.");
                 var response = await _profileLogic.AddToolToProfiles(name, profiles);
@@ -193,7 +198,8 @@ namespace IntelligenceHub.Controllers
         {
             try
             {
-                await GetTenantIdAsync();
+                var tenantResult = await SetUserTenantContextAsync();
+                if (!tenantResult.IsSuccess) return StatusCode(StatusCodes.Status500InternalServerError, tenantResult.ErrorMessage);
                 if (string.IsNullOrEmpty(name)) return BadRequest("Invalid request. Please check the route parameter for the profile name.");
                 if (profiles == null || profiles.Count < 1) return BadRequest("Invalid request. 'Profiles' property cannot be null or empty.");
                 var response = await _profileLogic.DeleteToolAssociations(name, profiles);
@@ -222,7 +228,8 @@ namespace IntelligenceHub.Controllers
         {
             try
             {
-                await GetTenantIdAsync();
+                var tenantResult = await SetUserTenantContextAsync();
+                if (!tenantResult.IsSuccess) return StatusCode(StatusCodes.Status500InternalServerError, tenantResult.ErrorMessage);
                 if (string.IsNullOrEmpty(name)) return BadRequest("Invalid request. Please check the route parameter for the profile name.");
                 var response = await _profileLogic.DeleteTool(name);
                 if (response.IsSuccess) return NoContent();
